Fix "multiple values for argument typemap" error
The problem was that the arguments to `from_datapackage` were messed up.
Due to a missing `join` call which caused a missing parentheses, the
string `'datapackage_facades/datapackage.json'` was treated as an
argument to `from_datapackage` namely as a value for the `typemap`
argument. Since this is then again supplied as a keyword argument, the
error was totally understandable.
I mean... isn't that totally obvious?